style: code review

+++ sequence-core/src/test/java/com/power4j/kit/seq/persistent/MySqlSeqHolderTest.java
@@ -45,7 +45,9 @@ public class MySqlSeqHolderTest extends SeqHolderTestCase {
 
 	@After
 	public void tearDown() {
-		seqSynchronizer.dropTable();
+		if (seqSynchronizer != null) {
+			seqSynchronizer.dropTable();
+		}
 	}
 
 	@Override
