+++ sequence-core/src/test/java/com/power4j/kit/seq/persistent/PostgreSqlSeqHolderTest.java
@@ -50,7 +50,9 @@ public class PostgreSqlSeqHolderTest extends SeqHolderTestCase {
 
 	@After
 	public void tearDown() {
-		seqSynchronizer.dropTable();
+		if (seqSynchronizer != null) {
+			seqSynchronizer.dropTable();
+		}
 	}
 
 	@Override
